state/python/requirements.txt
dapr==1.16.0
fastapi==0.111.0
grpcio==1.63.0
pydantic==2.4.2
requests>=2.32.4
uvicorn==0.23.2
certifi>=2024.02.02

